Set cmake-ide-* variables as defcustom; Mark safe
The cmake-ide variables at the top of the camke-ide file were defined
as global variables. These variables are not intended to be customized
by the user of this package. Also emacs displays a warning when these
variables are set in the .dir_locals file. This change defines these
variables as customizable variables and marked them as safe.