Add build pool and persistent naming
Currently cmake-ide requires a configured build directory per project,
if the build directory is to be reused.

This introduces the concept of a build pool directory, where all build
directories are created. If it is not configured, the system's tmp is
used (like previously done). This has the advantage that if tmp is size
limited (e.g. tmpfs), the user can specify a different global directory
altogether for cmake-ide to use.

This also introduces the concept of persistent naming where the
project's path is used as a basis for the build directory's name which
is predictable and reproducible, making it possible to reuse the build
directory across different editing sessions.

If both are used, the user can do away with the need to configure a
build directory per project.