Add support for setting flycheck-cppcheck-standards variable.
Previously, standards information was extracted from compiler flags
and passed to flycheck through the flycheck-clang-language-standard
variable. This commit modifies the relevant code to also pass
standards information to flycheck-cppcheck, through the
flycheck-cppcheck-standards variable.

cppcheck does not accept as many standards as clang/gcc/etc. do. The
default behaviour of the new code is to decay extracted standards into
the closest supported by cppcheck, where reasonable. If this fuzziness
is not desired, the (new) customisable variable
cmake-ide-cppcheck-strict-standards can be set non-nil; no information
will then be passed to flycheck-cppcheck unless it exactly matches
that extracted from the compile database.

Also included in this commit are two straightforward worker functions,
cmake-ide--valid-cppcheck-standard-p and
cmake-ide--cmake-standard-to-cppcheck-standard. Their behaviour is
consistent with the above.