Add unit tests for flycheck-cppcheck-standard functionality.
Added three unit tests. The two new worker functions
(cmake-ide--valid-cppcheck-standard-p and
cmake-ide--cmake-standard-to-cppcheck-standard) have a dedicated test
case each, and are tested as exhaustively as possible. The third test
case checks that the changes to cmake-ide-set-compiler-flags produce
the desired behaviour, whether or not the
cmake-ide-flycheck-cppcheck-strict-standards switch is set.

Also, one extra conversion pair is added (for c++0x -> c++03).