Fix bad insertion of sys-includes into company-c-headers-path-system.
This is a bug fix for Issue #108 on github. Previously, after a call
to cmake-ide-set-compiler-flags, the company-c-headers-path-system
variable would have the provided list of sys-includes consed on the
front as a nested list, causing company-c-headers to break. This
commit changes behaviour such that company-c-headers is modified to
simply append the sys-includes instead.

A regression test for this fix is included in the commit. Also, the
Cask file is updated to indicate a dependency on company-c-headers,
and cmake-ide-test.el now requires the same package.