Fix cmake invocation if build exists in project-dir
According to 'man cmake', the first non-option argument is ambiguous:
"{<path-to-source> | <path-to-existing-build>}"

If a build is already present in project-dir (yes, bad practice), then
cmake will use that as build directory (instead of cwd). As a result,
compile_commands.json ends up in project-dir, not where we look for it.